This adds rough support for parenting sections.
in the simple case this means if you declare a page section it has a parent_page acessor populated with the page containing it.
The new external_page_section let's you include a section that's logically part of another section but separated in the DOM.
Likewise, when elements are declared they can include a root option to break out of page section encapsulation.